+++ CHANGELOG.md
@@ -84,3 +84,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 ### Fixed
+- adjusted contact form message on mobile