+++ CHANGELOG.md
@@ -58,6 +58,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Contact us cards section to contact page
 - Linkedin link to footer
 - Next.js 13
+- .js files to house styles for all components
 
 ### Fixed
 
@@ -78,6 +79,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - next.js warning - no stylesheets in head component
 - CardColumns refactoring to accept an array of card objects as props
 - styles on newsletter button and contact page
+- error where html loaded before styles by updating \_document.js
+
+### Changed
+
+- all components to use styled components to handle all styling.
+
+### Removed
+
+- all .scss files with the exception of variables, mixins, and globals
 
 ## Unreleased